refactor: instanceof 분기 제거
