+++ src/main/java/domain/game/GameCommand.java
@@ -1,0 +1,7 @@
+package domain.game;
+
+import dao.GameDAO;
+
+public interface GameCommand {
+    void execute(Game game, GameDAO gameDAO, int gameId);
+}
