+++ src/main/java/domain/game/MoveCommand.java
@@ -1,11 +1,12 @@
-package view;
+package domain.game;
 
+import dao.GameDAO;
 import domain.board.Position;
-import domain.game.Game;
 
 public record MoveCommand(Position from, Position to) implements GameCommand {
     @Override
-    public void execute(Game game) {
+    public void execute(Game game, GameDAO gameDAO, int gameId) {
         game.move(from(), to());
+        gameDAO.updateMove(gameId, game, from(), to());
     }
 }
