--- src/main/java/view/GameCommand.java
@@ -1,7 +1,0 @@
-package view;
-
-import domain.game.Game;
-
-public interface GameCommand {
-    void execute(Game game);
-}
