--- src/main/java/view/PassCommand.java
@@ -1,10 +1,0 @@
-package view;
-
-import domain.game.Game;
-
-public record PassCommand() implements GameCommand {
-    @Override
-    public void execute(Game game) {
-        game.passTurn();
-    }
-}
